refactor(test): extract bundler env helpers into shared support module
Move `bundler_contaminated_env` and `overmind_invocation_log` from
dev_script_spec and dev_update_spec into spec/support/overmind_env_helpers.rb
to avoid divergence and make future updates to the env var list safer.
