+++ spec/lib/dev_update_spec.rb
@@ -7,9 +7,11 @@ require "socket"
 require "tmpdir"
 require "timeout"
 require_relative "../support/exec_tmpdir"
+require_relative "../support/overmind_env_helpers"
 
 RSpec.describe "bin/dev-update" do # rubocop:disable RSpec/DescribeClass
   include ExecTmpdir
+  include OvermindEnvHelpers
   let(:script_source) { File.expand_path("../../bin/dev-update", __dir__) }
   let(:poll_env) do
     {
@@ -183,24 +185,6 @@ RSpec.describe "bin/dev-update" do # rubocop:disable RSpec/DescribeClass
   def write_executable(path, contents)
     File.write(path, contents)
     FileUtils.chmod("+x", path)
-  end
-
-  def bundler_contaminated_env(dir)
-    {
-      "PATH" => "#{File.join(dir, 'stubbin')}:#{ENV.fetch('PATH')}",
-      "OVERMIND_SOCKET" => ".overmind.sock",
-      "BUNDLE_GEMFILE" => File.join(dir, "Gemfile"),
-      "BUNDLE_BIN_PATH" => "/tmp/fake-bundle-bin",
-      "BUNDLER_SETUP" => "/tmp/fake-bundler-setup",
-      "BUNDLER_VERSION" => "2.7.2",
-      "RUBYLIB" => "/tmp/fake-rubylib",
-      "RUBYOPT" => "-r/tmp/fake-bundler/setup",
-      "RUBYGEMS_GEMDEPS" => "-"
-    }
-  end
-
-  def overmind_invocation_log(dir)
-    File.read(File.join(dir, "stubbin", "overmind-env.log"))
   end
 
   def read_updater_log(dir)
